silence warnings in pancakes.swift
Hi! A friend just shared your tweet for this with me and I was eager to check out the actual code—that dynamic callable was pretty clever; couldn't work that part out just from looking at it. I did, however, notice that your two `is` checks emit warnings that they never fail, so I figured I'd PR a little something to change that :)